Implemented Quantization transform.
Matches colors to a quantized RGB colorcube of N levels
(for example, for GBA, N=32; For Amiga, N=16.),
with the option of using a 'clipped' mapping where
the brightest intensity level is not 1.0 but rather ((N-1)/N).

This kind of mapping was used for consoles that displayed on a CRT TV
-- eg SNES, Genesis, SMS -- and consequently had to avoid 'hot colors'
like pure white.

Mappings where N is different between components -- eg.
SVGA Hicolor 16bpp (32 levels of R, 64 of G, 32 of B),
or web palette (4R 8G 4B), are currently unsupported.